+++ packages/silicaui-charts/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @wizeworks/silicaui-charts
 
+## 0.5.0
+
 ## 0.4.0