+++ packages/silicaui-dnd/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @wizeworks/silicaui-dnd
 
+## 0.5.0
+
 ## 0.4.0