+++ packages/silicaui-table/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @wizeworks/silicaui-table
 
+## 0.5.0
+
 ## 0.4.0